move fetchAndProcessGivethProjects to index
